Adopt Servers & Tools pane in the add-in settings
